chore: update examples and document props
