--- examples/nextjs/src/pages/_app.js
@@ -1,12 +1,0 @@
-import {ThemeProvider, BaseStyles} from '@primer/react'
-import React from 'react'
-
-export default function App({Component, pageProps}) {
-  return (
-    <ThemeProvider>
-      <BaseStyles>
-        <Component {...pageProps} />
-      </BaseStyles>
-    </ThemeProvider>
-  )
-}
