+++ examples/nextjs/src/pages/index.tsx
